move fee from info api
