+++ api/info/client.go
@@ -27,7 +27,6 @@ type Client interface {
 	GetBlockchainID(context.Context, string, ...rpc.Option) (ids.ID, error)
 	Peers(context.Context, []ids.NodeID, ...rpc.Option) ([]Peer, error)
 	IsBootstrapped(context.Context, string, ...rpc.Option) (bool, error)
-	GetTxFee(context.Context, ...rpc.Option) (*GetTxFeeResponse, error)
 	Upgrades(context.Context, ...rpc.Option) (*upgrade.Config, error)
 	Uptime(context.Context, ...rpc.Option) (*UptimeResponse, error)
 	GetVMs(context.Context, ...rpc.Option) (map[ids.ID][]string, error)
@@ -91,18 +90,13 @@ func (c *client) Peers(ctx context.Context, nodeIDs []ids.NodeID, options ...rpc
 	return res.Peers, err
 }
 
+
 func (c *client) IsBootstrapped(ctx context.Context, chainID string, options ...rpc.Option) (bool, error) {
 	res := &IsBootstrappedResponse{}
 	err := c.requester.SendRequest(ctx, "info.isBootstrapped", &IsBootstrappedArgs{
 		Chain: chainID,
 	}, res, options...)
 	return res.IsBootstrapped, err
-}
-
-func (c *client) GetTxFee(ctx context.Context, options ...rpc.Option) (*GetTxFeeResponse, error) {
-	res := &GetTxFeeResponse{}
-	err := c.requester.SendRequest(ctx, "info.getTxFee", struct{}{}, res, options...)
-	return res, err
 }
 
 func (c *client) Upgrades(ctx context.Context, options ...rpc.Option) (*upgrade.Config, error) {
